stock-market-analysis-ai/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pandas==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
numpy==1.24.3
plotly==5.15.0
httpx==0.24.1
nltk==3.8.1
yfinance==0.2.18
groq